Decoupled reading and question activites.
