+++ prisma/seed.ts
@@ -158,36 +158,36 @@ const getAssignments = (courses: Course[]): Prisma.AssignmentCreateInput[] => [
 const getActivities = (assignments: Assignment[]): Prisma.ActivityCreateInput[] => {
   let activities = []
   for (let i = 0; i < assignments.length; i++) {
-    activities.push(
-      {
-        id: generateUUID(),
-        index: 0,
-        name: "Pretest",
-        description: `Pretest description`,
-        type: ActivityType.QUESTIONING,
-        assignment: { connect: { id: assignments[i]?.id }},
-      }
-    )
     // activities.push(
     //   {
     //     id: generateUUID(),
-    //     index: 1,
-    //     name: `Activity`,
-    //     description: `Activity Description`,
-    //     type: ActivityType.READING,
+    //     index: 0,
+    //     name: "Pretest",
+    //     description: `Pretest description`,
+    //     type: ActivityType.QUESTIONING,
     //     assignment: { connect: { id: assignments[i]?.id }},
     //   }
     // )
     activities.push(
       {
         id: generateUUID(),
-        index: 2,
-        name: `Posttest`,
-        description: `Posttest description`,
-        type: ActivityType.QUESTIONING,
+        index: 1,
+        name: `Activity`,
+        description: `Activity Description`,
+        type: ActivityType.READING,
         assignment: { connect: { id: assignments[i]?.id }},
       }
     )
+    // activities.push(
+    //   {
+    //     id: generateUUID(),
+    //     index: 2,
+    //     name: `Posttest`,
+    //     description: `Posttest description`,
+    //     type: ActivityType.QUESTIONING,
+    //     assignment: { connect: { id: assignments[i]?.id }},
+    //   }
+    // )
   }
   return activities
 }
